Minor change in Logging message.
